+++ junit5/src/main/java/com/baeldung/junit5/Greetings.java
@@ -1,4 +1,4 @@
-package com.baeldung.junit5.mockito;
+package com.baeldung.junit5;
 
 public class Greetings {
